+++ src/App.scss
@@ -52,7 +52,8 @@ $breakpoint-small: 460px;
 @import "scss/npc-options";
 @import "scss/npc-card";
 @import "scss/npc-stat";
-@import "scss/generate-npc-button";
+@import "scss/buttons";
+@import "scss/checkbox";
 
 @media(max-width: $breakpoint) {
 	.app__content {
