+++ src/components/npc-card.tsx
@@ -1,14 +1,5 @@
 import React, { Component } from 'react';
 import { Stat } from '../interfaces/stat';
-import { gods } from '../consts/gods';
-import { races } from '../consts/races';
-import { alignments } from '../consts/alignments';
-import { adjectives, traits, religiousAdjective, socioeconomic, traits2, quirks } from '../consts/npc-bio';
-import { randomNumber, getRandomNumberStandardDist, getRandomMapKey } from '../utility/functions';
-import { classes } from '../consts/consts';
-
-const baseUrl: string = 'https://www.dnd5eapi.co/api/';
-const abilities: string[] = ['strength', 'dexterity', 'constitution', 'intelligence', 'wisdom', 'charisma'];
 
 function NpcLineItem(props: any) {
 	return (
@@ -49,227 +40,6 @@ function NpcGenderIcon(props: any) {
 
 export class NpcCard extends Component<any, any> {
 
-	constructor(props: any) {
-		super(props);
-		this.state = {
-			name: null,
-			race: null,
-			class: null,
-			abilityScores: {
-				strength: { score: null, modifier: null },
-				dexterity: { score: null, modifier: null },
-				intelligence: { score: null, modifier: null },
-				wisdom: { score: null, modifier: null },
-				constitution: { score: null, modifier: null },
-				charisma: { score: null, modifier: null },
-			},
-			description: null,
-			attributes: {
-				age: null,
-				height: null,
-				weight: null,
-				alignment: null,
-				gender: null
-			},
-			loaded: false,
-			userSelections: {
-				alignment: null,
-				class: null,
-				gender: null,
-				level: null,
-				race: null
-			}
-		};
-	}
-
-	getNpcGender(): string {
-		return Math.floor(Math.random() * 2) === 0 ? 'Male' : 'Female';
-	}
-
-	getArmorClass(score: number): number {
-		return 10 + this.getModifier(score);
-	}
-
-	getHitpoints(hitDie: number, score: number, level?: number): number {
-		let hp = hitDie + this.getModifier(score),
-			rollsLeft = level || 0;
-
-		while (rollsLeft > 1) {
-			// TODO The modifier should adjust as con increases but it doesn't right now
-			hp += (randomNumber(1, hitDie) + this.getModifier(score));
-			rollsLeft--;
-		}
-
-		return hp;
-	}
-
-	getHeight(): string {
-		const height = getRandomNumberStandardDist(races.get(this.state.race.index)!.height.min, races.get(this.state.race.index)!.height.max, 1),
-			feet = Math.floor(height / 12),
-			inches = height % 12;
-
-		return feet + "' " + inches + '"';
-	}
-
-	getWeight(): number {
-		return getRandomNumberStandardDist(races.get(this.state.race.index)!.weight.min, races.get(this.state.race.index)!.weight.max, 1);
-	}
-
-	getNpcName(race: string, gender: string): string {
-		if (gender !== 'male' && gender !== 'female') {
-			gender = this.getNpcGender();
-		}
-		const rand1 = Math.floor(Math.random() * races.get(race)!.names.get(gender.toLowerCase())!.length),
-			rand2 = Math.floor(Math.random() * races.get(race)!.names.get('surname')!.length);
-
-		let last = races.get(race)!.names.get('surname')![rand2],
-			first = races.get(race)!.names.get(gender.toLowerCase())![rand1];
-
-
-		let returnName = first;
-		if (last) { returnName = returnName + " " + last; }
-
-		return returnName;
-	}
-
-	getModifier(score: number): number {
-		return Math.floor((score - 10) / 2);
-	}
-
-	generateStats(hitDie: number, abilityBonuses: any, numOfAbi: number, level: number): void {
-		let stats: { [key: string]: {} } = {},
-			ac: number = 10,
-			hitpoints = 0,
-			bonuses: Map<string, number> = new Map<string, number>([
-				['str', 0],
-				['dex', 0],
-				['con', 0],
-				['int', 0],
-				['wis', 0],
-				['cha', 0],
-			]),
-			availableAbi = numOfAbi * 2;
-
-		// If there are Ability bonuses, loop through and set them to the bonus map
-		if (abilityBonuses) {
-			abilityBonuses.forEach((ab: any) => {
-				bonuses.set(ab.ability_score.index, ab.bonus);
-			});
-		}
-
-		// Distribute level-based bonuses to bonus map
-		while (availableAbi > 0) {
-			let randomStat = abilities[randomNumber(0, abilities.length - 1)].substring(0, 3);
-			bonuses.set(randomStat, bonuses.get(randomStat)! + 1);
-			availableAbi--;
-		}
-
-		abilities.forEach((key: string) => {
-			let score: number = this.generateStat(),
-				keyTrunc: string = key.substring(0, 3);
-
-			if (bonuses && bonuses.has(keyTrunc)) {
-				// @ts-ignore
-				score += bonuses.get(keyTrunc);
-			}
-
-			if (key === 'dexterity') { ac = this.getArmorClass(score); }
-			if (key === 'constitution') { hitpoints = this.getHitpoints(hitDie, score, level); }
-			stats[key] = { score: score, modifier: this.getModifier(score) };
-		});
-
-		this.setState({
-			abilityScores: stats,
-			armorClass: ac,
-			hitpoints: hitpoints
-		});
-	}
-
-	generateStat(): number {
-		const reducer = (total: number, i: number) => total + i;
-		let statArray: number[] = [];
-
-		// 4d6 drop lowest
-		for (let i = 0; i < 4; i++) {
-			statArray.push(randomNumber(1, 6));
-		}
-
-		let min = Math.min(...statArray);
-		let index = statArray.indexOf(min);
-
-		statArray.splice(index, 1);
-		return statArray.reduce(reducer)
-
-		// Random
-		// return randomNumber(3,18);
-	}
-
-	generateDescription = () => {
-		const npc = this.state;
-		return (
-			<>
-				<p>{npc.name} is a {adjectives[randomNumber(0, adjectives.length)]} {npc.race.name} {traits[randomNumber(0, traits.length)]}</p>
-				<p>They {religiousAdjective[randomNumber(0, religiousAdjective.length)]} follower of {gods.get(npc.alignment.abbreviation)![0]} and {socioeconomic[randomNumber(0, socioeconomic.length)]} {traits2[randomNumber(0, traits2.length)]}</p>
-				<p>{quirks[randomNumber(0, quirks.length)]}</p>
-			</>
-		)
-	}
-
-	getAge = (): number => {
-		return getRandomNumberStandardDist(races.get(this.state.race.index)!.age.min, races.get(this.state.race.index)!.age.max, 1)
-	}
-
-	generateNpc = () => {
-		const selections = this.state.userSelections;
-		let randomClass = selections.class || classes[randomNumber(0, classes.length)];
-		let randomRace = selections.race || getRandomMapKey(races);
-		let randomAlignment = selections.alignment || getRandomMapKey(alignments)
-		let level = selections.level || randomNumber(1, 20);
-		let npcRace = this.getNpcData('race', 'races/' + randomRace);
-		let npcClass = this.getNpcData('class', 'classes/' + randomClass);
-		let npcLevel = this.getNpcData('level', 'classes/' + randomClass + '/levels/' + level);
-		let npcAlignment = this.getNpcData('alignment', 'alignments/' + randomAlignment);
-
-		Promise.all([npcRace, npcClass, npcLevel, npcAlignment]).then((r) => {
-			this.generateStats(r[1].class.hit_die, r[0].race.ability_bonuses, r[2].level.ability_score_bonuses, level);
-			const gender = selections.gender || this.getNpcGender();
-
-			this.setState({
-				name: this.getNpcName(r[0].race.index, gender),
-				attributes: {
-					age: this.getAge(),
-					height: this.getHeight(),
-					weight: this.getWeight(),
-					alignment: r[3].alignment,
-					gender: gender
-				}
-			}, () => {
-				this.setState({
-					description: this.generateDescription()
-				})
-			});
-
-			setTimeout(() => {
-				this.setState({ loaded: true });
-			}, randomNumber(1000, 3000));
-		});
-	}
-
-	async getNpcData(key: string, endpoint: string): Promise<any> {
-		try {
-			const response = await fetch(baseUrl + endpoint);
-			const data = await response.json();
-			this.setState({ [key]: data });
-			return { [key]: data };
-		} catch (error) {
-			this.setState({ [key]: 'problem' })
-		}
-	}
-
-	componentDidMount() {
-		this.generateNpc();
-	}
-
 	componentDidUpdate = (oldProps: any) => {
 		if (oldProps.userSelections !== this.props.userSelections) {
 			this.setState({ userSelections: this.props.userSelections });
@@ -288,14 +58,10 @@ export class NpcCard extends Component<any, any> {
 		return <NpcGenderIcon gender={gender}></NpcGenderIcon>
 	}
 
-	handleClick = (): void => {
-		this.setState({ loaded: false });
-		this.generateNpc();
-	}
-
 	render() {
-		const npc = this.state;
-		if (!npc.loaded) {
+		const npc = this.props.npcData,
+			loaded = this.props.loaded;
+		if (!loaded) {
 			return (
 				<>
 					<div className="npc-card"><div className="npc-card__spinner"></div></div>
@@ -304,13 +70,13 @@ export class NpcCard extends Component<any, any> {
 
 			)
 		}
-		console.log('NPC State', npc);
+		console.log('NPC State', this.props);
 		return (
 			<>
 				<div className="npc-card" >
 					<div className="npc-card__header">
 						<div className="npc-card__titles">
-							<h2 className="npc-card__name">{npc.name}{this.renderGenderIcon(this.state.attributes.gender)}</h2>
+							<h2 className="npc-card__name">{npc.name}{this.renderGenderIcon(npc.attributes.gender)}</h2>
 							<ol className="npc-card__info">
 								{this.renderLineItem(npc.race.name)}
 								{this.renderLineItem(npc.class.name)}
@@ -330,7 +96,7 @@ export class NpcCard extends Component<any, any> {
 							<li><strong>Age</strong>{npc.attributes.age} years</li>
 							<li><strong>Height</strong>{npc.attributes.height}</li>
 							<li><strong>Weight</strong>{npc.attributes.weight} lbs</li>
-							<li><strong>Alignment</strong>{npc.alignment.name}</li>
+							<li><strong>Alignment</strong>{npc.attributes.alignment.name}</li>
 						</ol>
 						<div className="npc-card__description">
 							{npc.description}
@@ -345,7 +111,6 @@ export class NpcCard extends Component<any, any> {
 						{this.renderAbilityScore('Charisma', npc.abilityScores.charisma)}
 					</ul>
 				</div>
-				<button className="generate-npc-button" onClick={this.handleClick}>Generate NPC</button>
 			</>
 		)
 	}
