+++ src/scss/_buttons.scss
@@ -1,0 +1,46 @@
+.button {
+	font-size: 1rem;
+	padding: 0.5rem 1.75rem;
+	margin-bottom: 1rem;
+	font-weight: 500;
+	color: var(--yellow);
+	background-color: var(--red);
+	border: 3px solid var(--red);
+	border-radius: var(--border-radius);
+	cursor: pointer;
+	transition: all 200ms ease;
+
+
+	&:hover {
+		transform: translateY(2px);
+		// box-shadow: 0 2px 6px 0px rgba(0, 0, 0, 0.3);
+	}
+
+	&:active,
+	&:focus {
+		color: var(--yellow);
+		background-color: var(--red);
+		border: 2px solid var(--red);
+	}
+
+	&--secondary {
+		font-size: 14px;
+		margin-right: 1rem;
+		color: var(--red);
+		background-color: transparent;
+		border-color: transparent;
+
+		&:hover {
+			color: var(--deep-red);
+			transform: none;
+		}
+	}
+
+	&--disabled {
+		background-color: gray;
+		border-color: gray;
+		color: darkgray;
+	}
+}
+
+@media (max-width: $breakpoint) {}
