+++ src/scss/_checkbox.scss
@@ -1,0 +1,53 @@
+.checkbox {
+	display: flex;
+	justify-content: flex-start;
+	align-items: center;
+	position: relative;
+	margin-bottom: 1rem;
+	cursor: pointer;
+
+	&__label {
+		display: inline-block;
+		color: var(--red);
+		font-size: 14px;
+		font-weight: 500;
+		padding-left: 0.75rem;
+	}
+
+	&__checkmark {
+		display: inline-block;
+		position: relative;
+		height: 1.25rem;
+		width: 1.25rem;
+		border-radius: var(--border-radius);
+		background-color: transparent;
+		border: 2px solid var(--red);
+
+		&:after {
+			position: absolute;
+			content: '';
+			display: none;
+			left: 5px;
+			top: 1px;
+			width: 4px;
+			height: 8px;
+			border: solid var(--white);
+			border-width: 0 3px 3px 0;
+			-webkit-transform: rotate(45deg);
+			-ms-transform: rotate(45deg);
+			transform: rotate(45deg);
+		}
+	}
+
+	&__hidden {
+		display: none;
+	}
+
+	&__hidden:checked~.checkbox__checkmark {
+		background-color: var(--red);
+
+		&:after {
+			display: block;
+		}
+	}
+}
