--- src/scss/_generate-npc-button.scss
@@ -1,41 +1,0 @@
-.generate-npc-button {
-	display: block;
-	font-size: 2rem;
-	padding: 1rem 3rem;
-	color: var(--yellow);
-	background-color: var(--red);
-	border: 2px solid var(--red);
-	margin: 0 auto;
-	box-shadow: var(--box-shadow);
-	border-radius: var(--border-radius);
-	cursor: pointer;
-	transition: all 200ms ease;
-
-	&:hover {
-		transform: translateY(2px);
-		box-shadow: 0 2px 6px 0px rgba(0, 0, 0, 0.3);
-	}
-
-	&:active,
-	&:focus {
-		color: var(--yellow);
-		background-color: var(--red);
-		border: 2px solid var(--red);
-	}
-
-	&--disabled {
-		background-color: gray;
-		border-color: gray;
-		color: darkgray;
-	}
-}
-
-@media (max-width: $breakpoint) {
-	.generate-npc-button {
-		position: fixed;
-		bottom: 0;
-		left: 0;
-		width: 100%;
-		box-shadow: 0px -4px 6px 0px rgba(0, 0, 0, 0.25);
-	}
-}
