+++ src/scss/_npc-options.scss
@@ -15,32 +15,20 @@
 	}
 
 	&__input {
-		flex: 1 1 auto;
+		flex: 1 1 200px;
 		margin: 0.5rem;
 
-		&--reset {
+		&--buttons {
+			display: flex;
+			justify-content: flex-end;
 			align-self: flex-end;
 			text-align: right;
 		}
 	}
 
-	&__button {
-		padding: 0.5rem 1.75rem;
-		margin-bottom: 1rem;
-		font-weight: 500;
-		color: var(--yellow);
-		background-color: var(--red);
-		border: 3px solid var(--red);
-		border-radius: var(--border-radius);
-		cursor: pointer;
-		transition: all 200ms ease;
+	//TODO break these into a partial with BEM names
 
-		&:hover {
-			transform: translateY(2px);
-		}
-	}
-
-	label {
+	label:not(.checkbox) {
 		display: block;
 		color: var(--red);
 		font-size: 14px;
@@ -62,74 +50,12 @@
 	}
 }
 
-.checkbox {
-	display: flex !important;
-	justify-content: flex-start;
-	align-items: center;
-	position: relative;
-	margin-bottom: 0.75rem;
-	cursor: pointer;
-
-	&__label {
-		display: inline-block;
-		color: var(--red);
-		font-size: 14px;
-		font-weight: 500;
-		padding-left: 0.75rem;
-	}
-
-	&__checkmark {
-		display: inline-block;
-		position: relative;
-		height: 1.25rem;
-		width: 1.25rem;
-		border-radius: var(--border-radius);
-		background-color: transparent;
-		border: 2px solid var(--red);
-
-		&:after {
-			position: absolute;
-			content: '';
-			display: none;
-			left: 5px;
-			top: 1px;
-			width: 4px;
-			height: 8px;
-			border: solid var(--white);
-			border-width: 0 3px 3px 0;
-			-webkit-transform: rotate(45deg);
-			-ms-transform: rotate(45deg);
-			transform: rotate(45deg);
-		}
-	}
-
-	&__hidden {
-		display: none;
-	}
-
-	&__hidden:checked~.checkbox__checkmark {
-		background-color: var(--red);
-
-		&:after {
-			display: block;
-		}
-	}
-}
-
 @media (max-width: $breakpoint) {
 	.npc-options {
 		flex-direction: column;
 
 		&__input {
 			flex: 1 1 100%;
-
-			&--reset {
-				text-align: right;
-			}
-		}
-
-		&__button {
-			font-size: 1.25rem;
 		}
 	}
 }
